+++ Sources/Sentry/include/SentryTransport.h
@@ -19,6 +19,10 @@ NS_SWIFT_NAME(Transport)
 
 - (void)recordLostEvent:(SentryDataCategory)category reason:(SentryDiscardReason)reason;
 
+- (void)recordLostEvent:(SentryDataCategory)category
+                 reason:(SentryDiscardReason)reason
+               quantity:(NSUInteger)quantity;
+
 - (SentryFlushResult)flush:(NSTimeInterval)timeout;
 
 #if defined(TEST) || defined(TESTCI) || defined(DEBUG)