+++ Tests/SentryTests/Networking/SentryHttpTransportTests.swift
@@ -2,6 +2,7 @@
 import SentryTestUtils
 import XCTest
 
+// swiftlint:disable file_length
 class SentryHttpTransportTests: XCTestCase {
 
     private static let dsnAsString = TestConstants.dsnAsString(username: "SentryHttpTransportTests")
@@ -10,7 +11,11 @@ class SentryHttpTransportTests: XCTestCase {
         let event: Event
         let eventEnvelope: SentryEnvelope
         let eventRequest: SentryNSURLRequest
+        let transaction: Transaction
+        let transactionEnvelope: SentryEnvelope
+        let transactionRequest: SentryNSURLRequest
         let attachmentEnvelopeItem: SentryEnvelopeItem
+        let transactionEnvelopeItem: SentryEnvelopeItem
         let eventWithAttachmentRequest: SentryNSURLRequest
         let eventWithSessionEnvelope: SentryEnvelope
         let eventWithSessionRequest: SentryNSURLRequest
@@ -53,15 +58,31 @@ class SentryHttpTransportTests: XCTestCase {
             event = Event()
             event.message = SentryMessage(formatted: "Some message")
 
+            let tracer = SentryTracer(transactionContext: TransactionContext(name: "SomeTransaction", operation: "SomeOperation"), hub: nil)
+            transaction = Transaction(
+                trace: tracer,
+                children: [
+                    tracer.startChild(operation: "child1"),
+                    tracer.startChild(operation: "child2"),
+                    tracer.startChild(operation: "child3")
+                ]
+            )
+            
             eventRequest = buildRequest(SentryEnvelope(event: event))
+            transactionRequest = buildRequest(SentryEnvelope(event: transaction))
             
             attachmentEnvelopeItem = SentryEnvelopeItem(attachment: TestData.dataAttachment, maxAttachmentSize: 5 * 1_024 * 1_024)!
-
+            transactionEnvelopeItem = SentryEnvelopeItem(event: transaction)
+            
             eventEnvelope = SentryEnvelope(id: event.eventId, items: [SentryEnvelopeItem(event: event), attachmentEnvelopeItem])
             // We are comparing byte data and the `sentAt` header is also set in the transport, so we also need them here in the expected envelope.
             eventEnvelope.header.sentAt = SentryDependencyContainer.sharedInstance().dateProvider.date()
             eventWithAttachmentRequest = buildRequest(eventEnvelope)
-
+            
+            transactionEnvelope = SentryEnvelope(id: transaction.eventId, items: [SentryEnvelopeItem(event: transaction), attachmentEnvelopeItem])
+            // We are comparing byte data and the `sentAt` header is also set in the transport, so we also need them here in the expected envelope.
+            transactionEnvelope.header.sentAt = SentryDependencyContainer.sharedInstance().dateProvider.date()
+            
             session = SentrySession(releaseName: "2.0.1", distinctId: "some-id")
             sessionEnvelope = SentryEnvelope(id: nil, singleItem: SentryEnvelopeItem(session: session))
             sessionEnvelope.header.sentAt = SentryDependencyContainer.sharedInstance().dateProvider.date()
@@ -530,6 +551,35 @@ class SentryHttpTransportTests: XCTestCase {
         XCTAssertEqual(clientReportRequest.httpBody, actualEventRequest?.httpBody, "Client report not sent.")
     }
     
+    func testTransactionRateLimited_RecordsLostSpans() {
+        let clientReport = SentryClientReport(
+            discardedEvents: [
+                SentryDiscardedEvent(reason: .rateLimitBackoff, category: .transaction, quantity: 1),
+                SentryDiscardedEvent(reason: .rateLimitBackoff, category: .span, quantity: 4)
+            ]
+        )
+        
+        let clientReportEnvelopeItems = [
+            fixture.attachmentEnvelopeItem,
+            SentryEnvelopeItem(clientReport: clientReport)
+        ]
+        
+        let clientReportEnvelope = SentryEnvelope(id: fixture.transaction.eventId, items: clientReportEnvelopeItems)
+        clientReportEnvelope.header.sentAt = SentryDependencyContainer.sharedInstance().dateProvider.date()
+        let clientReportRequest = SentryHttpTransportTests.buildRequest(clientReportEnvelope)
+        
+        givenRateLimitResponse(forCategory: "transaction")
+        
+        sut.send(envelope: fixture.transactionEnvelope)
+        waitForAllRequests()
+        
+        sut.send(envelope: fixture.transactionEnvelope)
+        waitForAllRequests()
+        
+        let actualEventRequest = fixture.requestManager.requests.last
+        XCTAssertEqual(clientReportRequest.httpBody, actualEventRequest?.httpBody, "Client report not sent.")
+    }
+    
     func testCacheFull_RecordsLostEvent() {
         givenNoInternetConnection()
         for _ in 0...fixture.options.maxCacheItems {
@@ -545,6 +595,26 @@ class SentryHttpTransportTests: XCTestCase {
         let deletedError = dict?["error:cache_overflow"]
         let attachment = dict?["attachment:cache_overflow"]
         XCTAssertEqual(1, deletedError?.quantity)
+        XCTAssertEqual(1, attachment?.quantity)
+    }
+    
+    func testCacheFull_RecordsLostSpans() {
+        givenNoInternetConnection()
+        for _ in 0...fixture.options.maxCacheItems {
+            sut.send(envelope: fixture.transactionEnvelope)
+        }
+        
+        waitForAllRequests()
+        
+        let dict = Dynamic(sut).discardedEvents.asDictionary as? [String: SentryDiscardedEvent]
+        XCTAssertNotNil(dict)
+        XCTAssertEqual(3, dict?.count)
+        
+        let transaction = dict?["transaction:cache_overflow"]
+        let span = dict?["span:cache_overflow"]
+        let attachment = dict?["attachment:cache_overflow"]
+        XCTAssertEqual(1, transaction?.quantity)
+        XCTAssertEqual(4, span?.quantity)
         XCTAssertEqual(1, attachment?.quantity)
     }
 
@@ -615,6 +685,29 @@ class SentryHttpTransportTests: XCTestCase {
         let dict = Dynamic(sut).discardedEvents.asDictionary as? [String: SentryDiscardedEvent]
         XCTAssertNotNil(dict)
         XCTAssertEqual(1, dict?.count)
+        
+        let attachment = dict?["attachment:network_error"]
+        XCTAssertEqual(1, attachment?.quantity)
+        
+        assertEnvelopesStored(envelopeCount: 0)
+        assertRequestsSent(requestCount: 1)
+    }
+    
+    func testBuildingRequestFails_RecordsLostSpans() {
+        sendTransaction()
+        
+        fixture.requestBuilder.shouldFailWithError = true
+        sendTransaction()
+        
+        let dict = Dynamic(sut).discardedEvents.asDictionary as? [String: SentryDiscardedEvent]
+        XCTAssertNotNil(dict)
+        XCTAssertEqual(3, dict?.count)
+        
+        let transaction = dict?["transaction:network_error"]
+        XCTAssertEqual(1, transaction?.quantity)
+        
+        let span = dict?["span:network_error"]
+        XCTAssertEqual(4, span?.quantity)
         
         let attachment = dict?["attachment:network_error"]
         XCTAssertEqual(1, attachment?.quantity)
@@ -931,6 +1024,15 @@ class SentryHttpTransportTests: XCTestCase {
     private func sendEventAsync() {
         sut.send(envelope: fixture.eventEnvelope)
     }
+    
+    private func sendTransaction() {
+        sendTransactionAsync()
+        waitForAllRequests()
+    }
+    
+    private func sendTransactionAsync() {
+        sut.send(envelope: fixture.transactionEnvelope)
+    }
 
     private func sendEnvelope(envelope: SentryEnvelope = TestConstants.envelope) {
         sut.send(envelope: envelope)
@@ -981,3 +1083,4 @@ class SentryHttpTransportTests: XCTestCase {
         XCTAssertEqual(0, dict?.count)
     }
 }
+// swiftlint:enable file_length